chore: separate field wrapper components
