+++ src/components/field-wrapper/FieldWrapper.tsx
@@ -50,34 +50,3 @@ export const FieldWrapper: React.FC<FieldWrapperProps> = ({
 }
 
 FieldWrapper.displayName = 'FieldWrapper'
-
-type InlineFieldWrapperProps = {
-  css?: CSS
-  error?: string
-  label: string
-  required?: boolean
-}
-
-export const InlineFieldWrapper: React.FC<InlineFieldWrapperProps> = ({
-  children,
-  error,
-  css,
-  label,
-  required
-}) => {
-  return (
-    <Box css={css}>
-      <Label
-        css={{ display: 'flex', alignItems: 'center' }}
-        required={required}
-      >
-        <Box css={{ mr: '$2' }}>{children}</Box>
-
-        {label}
-      </Label>
-      {error && <ValidationError css={{ mt: '$1' }}>{error}</ValidationError>}
-    </Box>
-  )
-}
-
-InlineFieldWrapper.displayName = 'InlineFieldWrapper'
